--- a/asv_bench/benchmarks/join_merge.py
+++ b/asv_bench/benchmarks/join_merge.py
@@ -157,6 +157,12 @@ class JoinIndex:
     def time_left_outer_join_index(self):
         self.left.join(self.right, on="jim")
 
+    def time_join_left_empty(self):
+        self.left.iloc[:0].join(self.right, on="jim")
+
+    def time_join_right_empty(self):
+        self.left.join(self.right.iloc[:0], on="jim")
+
 
 class JoinNonUnique:
     # outer join of non-unique